% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ongoal-package.R
\docType{package}
\name{ongoal-package}
\alias{ongoal}
\alias{ongoal-package}
\title{ongoal: NHL Data for Shots on Goal}
\description{
Data for several Pittsburgh Penguins seasons on whether a shots was on goal or not.
}
\author{
\strong{Maintainer}: Max Kuhn \email{max@rstudio.com} (\href{https://orcid.org/0000-0003-2402-136X}{ORCID})

Other contributors:
\itemize{
  \item RStudio [copyright holder]
}

}
\keyword{internal}
