% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rink-plot.R
\name{plot_nhl_shots}
\alias{plot_nhl_shots}
\title{Plot a hockey rink and add data}
\usage{
plot_nhl_shots(x, group = on_goal, symbols = c(yes = 1, no = 4))
}
\arguments{
\item{group}{A factor column to color and shape the data points.}

\item{symbols}{A named vector of symbol numbers for the plot (or NULL).}

\item{a}{data frame of data (should include a column called \code{on_goal}).}
}
\value{
A ggplot object
}
\description{
Plot a hockey rink and add data
}
